Add optional LinkedIn nickname
Add optional LinkedIn nickname instead of simply showing the LinkedIn ID. If you are unfortunate like my friend you may have a LinkedIn ID that is in this form: johndoe-8675235.
This addition enables you to add a second parameter that can be displayed instad for a more aesthetic appearance in the header.